refactor(security): 보안 및 인증 관련 클래스 패키지 구조 변경
