+++ src/main/java/gift/security/JwtTokenProvider.java
@@ -1,15 +1,15 @@
-package gift.service;
+package gift.security;
 
 import gift.entity.Member;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
 import java.util.Date;
 
-@Service
-public class JwtTokenService {
+@Component
+public class JwtTokenProvider {
     @Value("${jwt.secret}")
     private String secretKey;
     @Value("${jwt.expiration}")
